requirements.txt
flask==3.1.3
flask-cors==6.0.2
google-genai==1.68.0
PyPDF2==3.0.1
scikit-learn==1.5.2
nltk==3.9.1
python-dotenv==1.2.2
gunicorn==23.0.0
